# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: VPS Hosting Platform API
  version: 1.0.0
  description: |
    Backend API for a VirtFusion-style VPS hosting platform. This document is
    the contract a future UI team implements against.

    All responses use a consistent envelope:
      * Success: `{"data": ...}`
      * Error:   `{"error": {"code": "...", "message": "..."}}`

    Authentication uses opaque bearer session tokens returned by `POST /v1/auth/login`.
    Long-running operations (VM provisioning, start, stop, reboot, delete) are
    asynchronous: the create/lifecycle endpoints return an `operation_id`, and
    clients poll `GET /v1/operations/{id}` until the status is terminal.
  contact:
    name: Platform Engineering
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/vms:
    get:
      tags: [VMs]
      summary: List the current user's VMs
      operationId: listVMs
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of VMs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/VM"
    post:
      tags: [VMs]
      summary: Request a new VM
      description: >
        VM creation is asynchronous. Resources are derived from the selected
        plan. An `Idempotency-Key` header makes retries safe: repeating the
        same request returns the same operation and never creates a second VM.
      operationId: createVM
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          schema:
            type: string
          description: Client-generated key for idempotent retries
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVMRequest"
            examples:
              example:
                value:
                  plan_id: 6b17cffa-8f80-441a-b7fb-b22a414f4cdc
                  network_id: 6b17cffa-8f80-441a-b7fb-b22a414f4cdc
                  image_id: 6b17cffa-8f80-441a-b7fb-b22a414f4cdc
                  hostname: web-01
                  ssh_keys:
                    - - "ssh-ed25519 AAAA..."
      responses:
        "202":
          description: Provisioning started
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    example:
                      operation_id: "4f6d1a2b-..."
                      vm_id: "3e7d9c5c-..."
                    properties:
                      operation_id:
                        type: string
                        format: uuid
                      vm_id:
                        type: string
                        format: uuid
        "400":
          $ref: "#/components/responses/ValidationError"
        "402":
          $ref: "#/components/responses/BillingRequired"
        "503":
          $ref: "#/components/responses/NoCapacity"
...
  /v1/vms/{id}/console:
    post:
      tags: [Console]
      summary: Obtain a one-time console token (VNC or serial)
      description: >
        The returned token authorizes a WebSocket connection to the console
        gateway at `websocket_url`. Tokens are single-use and expire quickly.
        The underlying VNC endpoint / serial PTY is never exposed to the
        client. Choose the console with `console_type` (`vnc` for a graphical
        console via noVNC, `serial` for a text console over a Unix socket).
      operationId: getVMConsole
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ResourceIdPath"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                console_type:
                  type: string
                  enum: [vnc, serial]
                  default: vnc
                  description: "vnc = graphical (noVNC), serial = text console"
      responses:
        "200":
          description: Console access
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ConsoleAccess"
        "400":
          description: Unsupported console_type
        "409":
          description: VM is not running
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Session token from POST /v1/auth/login
  parameters:
    ResourceIdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    BillingRequired:
      description: An active subscription is required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NoCapacity:
      description: No node has sufficient capacity
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      example:
        error:
          code: RESOURCE_NOT_FOUND
          message: resource not found
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              example: RESOURCE_NOT_FOUND
            message:
              type: string
...
    CreateVMRequest:
      type: object
      required: [plan_id, network_id, image_id]
      properties:
        plan_id:
          type: string
          format: uuid
          description: Selects the plan whose resources size the VM
        network_id:
          type: string
          format: uuid
        image_id:
          type: string
          format: uuid
        name:
          type: string
        hostname:
          type: string
        ssh_keys:
          type: array
          items:
            type: string
          description: SSH keys written to cloud-init
        root_password:
          type: string
          description: Optional cloud-init root password (not recommended)
        idempotency_key:
          type: string
          description: Alternative to the Idempotency-Key header
    VM:
      type: object
      example:
        data:
          id: "3e7d9c5c-..."
          user_id: "6b17cffa-..."
          name: web-01
          hostname: web-01
          status: running
          vcpu: 2
          memory_mb: 2048
          disk_gb: 40
          mac_address: 02:00:00:00:00:01
          created_at: "2026-09-08T03:00:00Z"
      properties:
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            user_id:
              type: string
              format: uuid
            plan_id:
              type: string
              format: uuid
              nullable: true
            node_id:
              type: string
              format: uuid
              nullable: true
            network_id:
              type: string
              format: uuid
              nullable: true
            name:
              type: string
            hostname:
              type: string
            status:
              type: string
              enum: [pending, provisioning, running, stopped, suspended, terminated, error]
            vcpu:
              type: integer
            memory_mb:
              type: integer
            disk_gb:
              type: integer
            mac_address:
              type: string
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...
    ConsoleAccess:
      type: object
      example:
        data:
          console_type: vnc
          token: "5xt9..."
          expires_at: "2026-09-08T03:02:00Z"
          websocket_url: "wss://panel.example.com/console/ws?token=5xt9..."
      properties:
        data:
          type: object
          properties:
            console_type:
              type: string
              enum: [vnc, serial]
            token:
              type: string
              description: Single-use token embedded in websocket_url
            expires_at:
              type: string
              format: date-time
            websocket_url:
              type: string
              format: uri
...
security:
  - bearerAuth: []